Fix Project Properties Exception
If the project properties has focus, the DTE.ExecuteCommand throws an
exception. To get around this, if it fails, the code now gets the
keybindings for Edit.Copy/Cut, and uses SendKeys auto-click those
keyboard keys. Seems to work if you leave Edit.Copy bound to Ctrl+Ins.
Did not test with alternate bindings.